Added validation for role name when modif/create and user email when create

+++ frontend/src/app/components/modify-role/modify-role.component.ts
@@ -1,5 +1,12 @@
 import { Component, OnInit } from '@angular/core';
-import { FormControl, FormGroup, Validators } from '@angular/forms';
+import {
+  AbstractControl,
+  AsyncValidatorFn,
+  FormControl,
+  FormGroup,
+  ValidationErrors,
+  Validators,
+} from '@angular/forms';
 import { RoleApiService } from '../../services/role-api.service';
 import { ActivatedRoute, Router, RouterModule } from '@angular/router';
 import { IconTextButtonComponent } from '../../shared/components/buttons/icon-text-button/icon-text-button.component';
@@ -9,6 +16,8 @@ import { LongInputComponent } from '../../shared/components/form/long-input/long
 import { PermissionsTreeComponent } from '../permissions-tree/permissions-tree.component';
 import { Role } from '../../models/Role';
 import { PermissionDirective } from '../../shared/directives/permission.directive';
+import { Observable, of, map, catchError, lastValueFrom } from 'rxjs';
+import { SnackbarService } from '../../shared/components/snackbar/snackbar.service';
 
 @Component({
   selector: 'app-modify-role',
@@ -32,7 +41,11 @@ export class ModifyRoleComponent implements OnInit {
 
   initialPermissions: Set<string> | undefined = undefined;
 
-  roleName = new FormControl('', Validators.required);
+  roleName = new FormControl('', {
+    validators: Validators.required,
+    asyncValidators: this.roleNameValidator(),
+    updateOn: 'blur',
+  });
   roleState = false;
   roleDescription = new FormControl('', Validators.maxLength(255));
   rolePermissions = new Set<string>();
@@ -46,6 +59,7 @@ export class ModifyRoleComponent implements OnInit {
     private api: RoleApiService,
     private router: Router,
     private route: ActivatedRoute,
+    private snackbar: SnackbarService,
   ) {}
 
   ngOnInit(): void {
@@ -62,23 +76,34 @@ export class ModifyRoleComponent implements OnInit {
   }
 
   save() {
+    const deactivatable = !this.roleState
+      ? lastValueFrom(this.api.isDeactivatable(this.role.id!))
+      : Promise.resolve(true);
     if (this.group.valid) {
-      this.api
-        .updateRole({
-          id: this.role.id!,
-          name: this.roleName.value!,
-          state: this.roleState,
-          description: this.roleDescription.value!,
-          permissions: Array.from(this.rolePermissions),
-        })
-        .subscribe(() => {
-          this.mode = 'consult';
-          this.initialPermissions = this.rolePermissions;
-          this.role.permissions = Array.from(this.rolePermissions);
-          this.role.name = this.roleName.value!;
-          this.role.state = this.roleState;
-          this.role.description = this.roleDescription.value!;
-        });
+      deactivatable.then((isDeactivatable) => {
+        if (isDeactivatable) {
+          this.api
+            .updateRole({
+              id: this.role.id!,
+              name: this.roleName.value!,
+              state: this.roleState,
+              description: this.roleDescription.value!,
+              permissions: Array.from(this.rolePermissions),
+            })
+            .subscribe(() => {
+              this.mode = 'consult';
+              this.initialPermissions = this.rolePermissions;
+              this.role.permissions = Array.from(this.rolePermissions);
+              this.role.name = this.roleName.value!;
+              this.role.state = this.roleState;
+              this.role.description = this.roleDescription.value!;
+            });
+        } else {
+          this.snackbar.show(
+            'This role cannot be deactivated, it is being used by some users',
+          );
+        }
+      });
     }
   }
 
@@ -88,6 +113,7 @@ export class ModifyRoleComponent implements OnInit {
 
   cancel() {
     this.roleName.setValue(this.role.name);
+    this.roleName.markAsUntouched();
     this.roleDescription.setValue(this.role.description);
     this.roleState = this.role.state;
     this.rolePermissions = this.initialPermissions!;
@@ -99,8 +125,31 @@ export class ModifyRoleComponent implements OnInit {
   }
 
   deleteRole() {
-    this.api.deleteRole(this.role.id!).subscribe(() => {
-      this.router.navigate(['/roles']);
+    this.api.isDeactivatable(this.role.id!).subscribe((isDeactivatable) => {
+      if (isDeactivatable) {
+        this.api.deleteRole(this.role.id!).subscribe(() => {
+          this.router.navigate(['/roles']);
+        });
+      } else {
+        this.snackbar.show(
+          'This role cannot be deleted, it is being used by some users',
+        );
+      }
     });
+  }
+
+  roleNameValidator(): AsyncValidatorFn {
+    return (control: AbstractControl): Observable<ValidationErrors | null> => {
+      if (!control.value) {
+        return of(null);
+      }
+
+      return this.api.validateRoleName(control.value).pipe(
+        map((isValid) =>
+          isValid || control.value == this.role.name ? null : { unique: true },
+        ),
+        catchError(() => of(null)), // Treat as valid if there's an issue with the request
+      );
+    };
   }
 }
